# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from './approvalStatus';

export interface Approval {
  id: number;
  runId: number;
  title: string;
  detail: string;
  status: ApprovalStatus;
  expiresAt: Date;
}