# lib/api-zod/src/generated/types/memoryNote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryNote {
  id: number;
  title: string;
  path: string;
  excerpt: string;
  scope: string;
  provenance: string;
}